refactor(infra): use MediaTypes.TextPlain constant and extract KeycloakTokenProvider (US-000)
Replace hardcoded "text/plain" in ShoppingEndpoints with MediaTypes.TextPlain
constant. Extract token acquisition logic from KeycloakAdminService (300 lines)
into KeycloakTokenProvider (~80 lines) with its own HttpClient registration.
KeycloakAdminService now focuses on user operations only (~230 lines).
