+++ src/Yumney.Shared/Common/MediaTypes.cs
@@ -6,4 +6,5 @@ public static class MediaTypes
     public const string ImagePng = "image/png";
     public const string ImageWebp = "image/webp";
     public const string TextEventStream = "text/event-stream";
+    public const string TextPlain = "text/plain";
 }
